[ftrace importer] Use tid instead of pid for kernel threads
Always use tids as keys when retrieving threads from the
threadsByLinuxPid object since a Linux Kernel PID is the
equivilant of a POSIX TID in userspace.

This prevents initializing a thread in the model with an incorrect
thread name if the the pid and tid do not match and threads within
a process are named differently.

A reproducer for the original issue that this change fixes is also
added as a new unit test for the ftrace importer.

Change-Id: Ifa7c72273412e8fb9c52a1e744e2834872ad67dc