Android systrace: Fix printf/scanf format compiler warnings
With some compiler toolchains the type uint64_t is defined as the
type "unsigned long long." On others, it is simply defined as
"unsigned long." In the later case this can result in compiler
warnings when attempting to print variables of type uint64_t with
the "%llu" printf format specifier.

This change switches atrace_helper to use the PRIu64 and and PRIx64
instead of "%ull" and "%xll" respectively when using printf/scanf.
This allows the atrace_helper compilation to be agnostic to the
compiler toolchain's definition of uint64_t.

This change allows atrace_helper to compile without warnings using
recent versions of clang such as the version that is included as
part of the current AOSP toolchain.

Change-Id: If9fef3acc8c608a6bd91b275cdd57df3aacd5ad4