Remove BattOr support from systrace
As decided in https://goo.gl/g5Srzn, the low number of regressions
caught by the power metrics alone doesn't justify the high cost of
maintaining the BattOrs on the waterfall or the Python code that
supports BattOr tracing.

After this CL is submitted, I'll remove the common code shared by
Telemetry and systrace.

Bug: chromium:859514
Change-Id: Ib72a2ef03be1773dcbc865f380b9bd968b603cba