[Telemetry] Add flush-tracing and sync markers on first tab only
The DevTools client might list contexts which do not yet have a live
DevTools instance to connect to (e.g. background tabs which may have
been discarded or not yet created).

In some situations trying to place markers on all contexts will
cause us to hang when we hit one with a non yet live DevTools
instance. To prevent this, and since this is enough for tracing
purposes, markers will now be added to the firstly created tab only.

Bug: chromium:854212
Change-Id: If897b7903102590a58224546eb713380849fd03d