Support Debian unstable
Debian have removed support for lsb-release. This means that there is
no lsb-package (and thus no /etc/lsb-release) in Debian unstable right
now. Telemetry requires that file to detect the name and version of
the Linux distro being used. Not having that file means that telemetry
cannot be run locally.

It would be nice to have some fallback. This CL adds support for the
fairly common /etc/os-release file which is widely available in many
Linux distros.

Bug: catapult:#
Change-Id: Iad0fcc98bf5256176c4024c8a85c7db8183966e9