Add DeleteAllEntities().
There are currently many entities in the datastore whose models have been or can
be removed from the source.
Attempts to access these entities raises KindError, which are occasionally seen
in the logs.
This CL adds a function that can be called from dev console to delete all
entities of a given kind.

Bug: 851127
Change-Id: I01c467cb0264ff0eadf36218c960ddf4ebaf1807