Update patch set 5
Patch Set 5:

I like the idea of being able to find the code that emits the trace events.

But since trace viewer is not used only by chrome (it is used by android as well as other external parties), adding an explicit chromium search to all events will be confusing to a lot of our users.

If this code can be moved to tracing/extras/chrome/ so that it can only appear for chrome traces, that would be great.

Also, please file an issue on the catapult issue tracker, and link it using catapult:XXXXXX.

Patch-set: 5
CC: Gerrit User 1115914 <1115914@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>