Call thunks directly within other thunks.
Taking from the Polymer Redux Tips document,
"Think carefully about when you dispatch an action. That causes polymer to
re-render, which can cost a lot of CPU time, which would be wasted if another
action is dispatched shortly after. Instead of dispatching multiple reducers in
rapid succession, you can either define a single reducer that calls other
reducers directly, or use a meta reducer like chain that takes several actions
and calls them directly through the reducer registry."

These changes only cover portions of the code that CL 1130105 did not completely
rewrite. There will be more of these changes in that CL, but this would be a
great opportunity to get some early feedback on this method.

It is also worthwhile to note that Redux DevTools no longer goes past the
default action limit. Everything action is now recorded from the very start of
the application. This seemed to speed up the application overall while using
Redux DevTools.

Bug: chromium:855252
Change-Id: I6df708fe11dbab5a6585463e41273992c98e9316