Benchmark IndexedDB in various scenarios.
IndexedDB is currently chosen to be the datastore behind timeseries
caching in v2spa. These performance benchmarks are designed to test
the various ways of implementing different internal data structures and
timing different methods for reading/writing data. The purpose of these
testbenches are to find fast ways of caching data in IndexedDB.

Bug: chromium:855252
Change-Id: I079646519d4a0b69f6c5fa2e0b02ea6c7165a3ff