Initial implementation
In big lines, the data was loaded from results.html ( the user is able to choose
any file using the button ). Regarding the fact that there are both code lines
and <histrogram-json>text</histrogram-json> tags in the same file, I extracted
the data between these tags.

There are 3 types of information depending on the fields of JSONs, so I divided
them in 3 structures (for some of them I created objects with specific fields).

Then I matched every story from one category with its guid from other category.

The table displays the name, the stories, the sampleValue, the value and the
storySetRepeats, given the fact that stories and values are linked by guid.

Bug: catapult:#
Change-Id: I5674ebebd762439c8e71033d887296bd0e7d4ea3