[pinpoint] Add ability to add aliases to bot configurations.
The alias just serves as an alternative name to look up a bot.
This will reduce configuration duplication and therefore reduce errors.
It also gives us a mechanism to differentiate between configurations
that are canonical, and appear in the bot configurations listing,
and aliases, which don't.

To add an alias, add a bot where the only field is the "alias" field.
{
  "alias-name": {"alias": "canonical-name"},
  "canonical-name": {"browser": "release", "dimensions": {...}}
}

Bug: catapult:#4561
Change-Id: Ib13c539ee367b9aedf4d2b1923a9d7df94253703