Fix descriptors for huge test suites.
Some test suites are so huge that, not only can their keys not fit in 1GB RAM,
but it also takes longer than 10 minutes to fetch and process them.
Also, sometimes the iterator runs for a few minutes and then throws an
inscrutable BadRequestError 'invalid handle'.
This CL runs the iterator for as long as possible, then defers another task to
pick up where it left off.

Bug: catapult:#4461
Change-Id: I13abde02764e12e42e3fbc5a107f5acb1252dd63