Update patch set 6
Patch Set 6:

(7 comments)

This looks great! It simplifies many of the steps that we perform in `element-base.js`. Very clean code and lots of documentation!

I do have some concerns about the direction this is heading towards, mainly the use of `registerReducer` and usage of "state path". Before I begin, I should state that I gravitate towards functional programming and my workflow for Redux came from my experience working in React, which might not always agree with how Polymer works.

The current state of registering reducers relies on extending from the base class, `ElementBase` then specifying actions and reducers in a static property before calling `ElementBase.registerReducers`. This couples Redux actions and reducers with the component, which became confusing for me while working with the code. Both Polymer and Redux have a notion of "dispatching"; Polymer dispatches events to the DOM and Redux dispatches actions to the store. A clear separation between Redux and Polymer would have helped, possibly through the use of ES6 modules and dedicated files for Redux logic.

 > This example project showcases one method of separating Polymer and Redux logic:
 > https://github.com/Polymer/pwa-starter-kit-hn/tree/master/src/components

Another concern this brings up is the one-to-one relationship between current state actions and reducers. Scenario: let's say that we wanted to render a loading indicator in the top appbar while waiting for the login response. Two actions would have to be dispatched: one to specify loading to the login page's reducer and another to the appbar's reducer. Now imagine if reducers could listen to every action: one loading action would be dispatched and picked up by both reducers. Allowing reducers to listen to multiple actions will decrease the amount of dispatches and allow for more powerful reducer behavior. With the introduction of the "CHAIN" action in this CL (good job!), it is possible to keep the amount of dispatches down; but, current state reducers are still limited to their own actions.

My apprehensiveness of "state path" comes mainly from working in React for 4 years, where it is largely unpopular. Instead, the community makes use of selectors to derive state:

  const selector = (state) => state.chartsById[lineId].lines;
  const path = `${rootPath}.chartsById.${lineId}.lines`;

This ultimately leads to less confusion and less reliance on magic. Selectors are explicit about what is selected. This becomes even more helpful if you are using TypeScript or Flow to provide typing to your Redux state.

 > This file showcases listening for multiple actions and creating selectors:
 > https://github.com/Polymer/pwa-starter-kit-hn/blob/master/src/reducers/items.js

Hopefully this sparks some discussion; I look forward to hearing any feedback!

Patch-set: 6
Reviewer: Gerrit User 1290511 <1290511@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>