Finish api change in preparation for integration.
This commit updates the API for graph data as discussed in
the meeting so that it is ready to integrate with Florina's
work for a demo version. To allow for the format change it also
pushes the assignment of colors to inside the graph data
class. It may make sense to assign ids to the
plots being drawn so that the caller can change the colors in CSS, but
for now that flexibility probably isn't required.

Bug: chromium:866423
Change-Id: Id998bbcc68dd61d3fdcf14cd839cf62f99fc75b5