Add new flag '--compatibility-mode' to Run Benchmark in Compatibility Mode
This flag is added and will be used in CL:1150032 and later CLs. So that we could run benchmarks in early versions of Chrome.
The reason for adding this benchmark is that earlier versions of Chrome (versions before M60) have a configuration file size limit, so we need to remove the field trials command using this flag to reduce file size, also, '--ignore-certificate-errors-spki-list' flag is not supported in earlier versions of the browsers, so we need to use '--ignore-certificate-errors' instead. There might be some other configuration change that is required in future, which can be made using this flag.

Bug: chromium:866881
Change-Id: I8d26fd8bd667b97faa0701c8063b5a4b641d590c