Update patch set 4
Patch Set 4:

I love it! My only negative feedback is that, when Annie showed me the demo, I wasn't sure what "Values" meant. I'm not sure if the correct route is to change it to "Values present", or to add a tooltip, or to just do nothing because I'm an idiot, but it definitely wasn't immediately obvious to me what a green box next to values meant.

Patch-set: 4
CC: Gerrit User 1128909 <1128909@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>