Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > This is great Dave, I love this!
> 
> I love it too!
> 
> > * I found it confusing to understand what happens when you click on a dot on the graph. I saw the commit's changing, but wan't sure exactly just how. After a while I realized that clicking on a dot selects that commit to appear on the right side panel, and the one on the left side  is always the immediately preceding commit. I think that's fine, but the confusion could be solved if:
> >   - somehow, there was some visual feedback on the dots when selecting them; i.e. highlight the two ones that are being shown below.
> 
> That was my first thought, too. I think the behavior of clicking on a dot making it show on the right side is the right thing, but helping users understand that's what's happening would be awesome.

Yes, I do plan on implementing this! Some of the early mocks of the revamped chart showed something similar to this: https://github.com/catapult-project/catapult/issues/3877#issuecomment-335289043

Those mocks were before I decided to do the side-by-side comparison. Now, I think we would show the chart in black, with the selected commit in the accent color (pink) and the previous commit in the base color (blue). And we can show the side histogram in the two colors as well. We can have colored visual markers with the left and right commits to show that they go with the colors on the chart.

It's difficult to do this kind of selection in flot, so I am blocking on rewriting the chart in d3.

Patch-set: 4
Reviewer: Gerrit User 1000093 <1000093@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>