Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Thanks Juan! I think this overlaps with https://chromium-review.googlesource.com/c/catapult/+/1152175 -- +Changwan to help sort it out.
> 
> Yeah, I only noted after uploading and noticing the "merge conflict". I'm fine whichever people prefer to land. They are nearly identical except for a couple of places where we diverged on how to silence a warning.

Thanks a lot for looking into the presubmit error, and writing this up!
They surely look identical. It's late hours in london and therefore it would make sense for me to land the other CL and follow up if necessary.

I'll make sure to cc you or update on the original CL as soon as possible to avoid this type of confusion.

Patch-set: 1