Refactor GraphPlotter to use strategy pattern.
The previous design of the graph plotter would make it difficult
to add new plots and would quickly lead to an overcomplicated class.
This helps to mitigate these issues by only leaving the code common
to all plots in graph plotter and exposing a plot method that which
accepts plotters for the specific graph types.

Bug: chromium:866423
Change-Id: Ibf3749d3b73091117c586a2a3a554832a08f2d00