[Review this First] Fix Bug for Compatibility Mode
This CL tries to fix the test error in CL:1150032. This change should be made
in CL:1150026, however I thought it would not be necessary since if you
call the `UpdateFromParseResults` function in browser_option.py, the field
`compatibility_mode` will be set. While in the unit test, this is not the case.
So I added the field `compatibility_mode` so that it will be initialized right
after the creation of BrowserOptions instance. And the code in CL:1150032 can
access this field in the test cases.


Bug: chromium:866881
Change-Id: I692ce881abd4074827c1d8959de116a4632ce504