Update patch set 7
Patch Set 7:

(2 comments)

Hi,

Here is the feature with the drop-down menu, but I also left the set of buttons which did the same thing. ( you can choose a metric + story and have the results ).

At this step, the final information from drop-down menu is not displayed, but it is built behind.

I think that all the tables & additional displaying features will be useless because the obtained data from buttons & drop-down menu should go straight in plots and histograms, so
it might be fine to remove some of them.

And sorry for the delay but it took me a while to accommodate with Vue.js.

Patch-set: 7
Reviewer: Gerrit User 1293425 <1293425@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>