Use Legacy Command Line Path in Compatibility Mode.
In this CL, we let the `FlagChanger` class to used the
`_CMDLINE_DIR_LEGACY` flag instead of the `_CMDLINE_DIR` if it's in
compatibility mode. The reason is that for milestone before 54, Chrome will
try to find the command line config file from the legacy path. The default
is to use `_CMDLINE_DIR`.

Correspondingly, we let the caller of the `FlagChanger` in
`android_browser_finder.py` to pass in the compatibility_mode flag.

There is also some minor style change in this CL.

Bug: chromium:866881
Change-Id: I5f45293cf83365636bd6aa080d7acb90a3f99833