Update patch set 7
Patch Set 7:

I haven't looked at in detail yet, just a couple of high-level questions.
> Patch Set 7:
> 
> (2 comments)
> 
> Hi,
> 
> Here is the feature with the drop-down menu, but I also left the set of buttons which did the same thing. ( you can choose a metric + story and have the results ).
> 
> At this step, the final information from drop-down menu is not displayed, but it is built behind.

Would adding a text box and displaying the filtered data as a string work? I know we would not need it when we integrate your work with Anthony's but test it might be easier. If this is too difficult, may be just a console.log would do as well. Just so we can test it before we integrate. 

> 
> I think that all the tables & additional displaying features will be useless because the obtained data from buttons & drop-down menu should go straight in plots and histograms, so
> it might be fine to remove some of them.

Yes, I think removing tables and support for building table could be removed. Do you want to remove them in a different cl? I am fine with removing in this cl or in a different cl.
> 
> And sorry for the delay but it took me a while to accommodate with Vue.js.

No problem at all  :)

Patch-set: 7
Reviewer: Gerrit User 1145189 <1145189@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>