Update patch set 6
Patch Set 6:

I've created and added to the CL description a bug about supporting Fuchsia devices with Catapult. As for the design doc, I'm still just trying to pull a minimal set of functionality from similar implementations of pieces of the telemetry codebase for CrOS and Android, and then modifying them to work for Fuchsia. I believe I've currently got it in a functional state, and I'm just filling in functions that throw NotImplementedErrors. I'm still not completely clear on control flow inside of Catapult, or on the expected object interfaces.

Patch-set: 6
Tag: mailMessageId=<CAAhdxuAPhmU0DTupYMbYgMs3MTZuK-kmUG8wsc+NWDm-cpMw3A@mail.gmail.com>