Update FindBrowser() doc-string and its browser chosen code.
browser_finder.FindBrowser() used to return None if no browser is
found. However it is inconsistent with its document. Given the fact
that many users of FindBrowser() already handle None instead of
BrowserFinderException, it is easier to fix document instead.

Also, rewrite browser picker code to make it more readable.

BUG=chromium:867328

Change-Id: Ie030a4fa1999a7caf2fbb06b5dc51db45ab79b82