Removed an old restriction on timestamp ordering of trace events
After encountering a "begin" event but before the corresponding "end"
event, the trace event importer would discard most thread-scoped events
that had a ts before the "begin" event. There's nothing in the tracing
code preventing this however, since timestamps can be manually specified.
As far as I can tell there's nothing relying on this behavior either,
as the slices in a slicegroup are sorted before toplevel slices are
computed.

R=benjhayden@chromium.org