Add tool for visualizing metrics times
This CL adds a button to results.html that appears when values from a
rendering benchmark is present. This button toggles the display of a
stacked column chart, which displays different metrics related to the
rendering pipeline. This chart also allows viewing further breakdowns of
the durations of each column.

More information on this tool can be found here:
https://docs.google.com/document/d/1bxDJKYJtcp-kmM3CdDkpv8SQpv8HwJOgBCj-1Z2JBfA/edit?usp=sharing

Bug: catapult:#4562
Change-Id: Ia65c3bf63a77724262d69c4650b23f4341a190c5