Update patch set 10
Patch Set 10:

> Patch Set 6:
> 
> I've created and added to the CL description a bug about supporting Fuchsia devices with Catapult. As for the design doc, I'm still just trying to pull a minimal set of functionality from similar implementations of pieces of the telemetry codebase for CrOS and Android, and then modifying them to work for Fuchsia. I believe I've currently got it in a functional state, and I'm just filling in functions that throw NotImplementedErrors. I'm still not completely clear on control flow inside of Catapult, or on the expected object interfaces.

This is looking good. But now it's probably a good time to start breaking this into smaller CLs, for example, each introducing a few "fuchsia" class and adding tests for it (even if at the beginning they are just unit tests with the actual fuchsia interface mocked out).

I also think some sort of "design doc" would be useful, even if just to track your work and indicate: which new classes need to be created (what are they analogs of), any refactors that would be needed on existing classes for other platforms, and the order in which you plan to land these things.

Patch-set: 10
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>