Update patch set 12
Patch Set 12:

Hi Leszek and Mythri,

After having a discussion with Mythri, we decided to change the way a diagnostic is chosen.

So instead of doing intersection of diagnostics and avoiding the rest of elements (because I noticed that common metric + common story don't mean common diagnostics fields ), I did keep all the elements. The user has the option to choose a diagnostic that is not common, but those elements which have not this type of diagnostic will be avoided.

Thank you!

Patch-set: 12
Reviewer: Gerrit User 1293425 <1293425@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>