Filtering and selection updates
There is a simulation for the future dropdown-menu in Vue.js.
There are 2 text boxes and the user can insert the metric + story.
Also, there is the drop-down menu and I think I have to remove
all part with tables and so on.

The partial results will be displayed: sample values, matched guid etc.
Also the partial result from the drop-down menu is not displayed
because this should go in histograms.

BUG=chromium:866423
Change-Id: I798010f635013f7b5398948afb2d0a588fdca3d3