Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> Thanks for working with the existing charts!
> It's been a while since I worked with d3, so I would really rather not review hundreds of lines of d3. I'd rather you add as little code as possible to the existing NameBar/ColumnChart classes.
> What exactly do you really need your charts to do that the existing NameBar/ColumnChart don't do?

The main issue is that the hover labels of the current charts don't display enough information. For instance, since the columns are grouped by story, the hover label should show the name of the story, as well as the display label. This helps the user to compare the metric values both across pages and between different runs of the same page. However, the only way to store this information is within the data of the chart itself, but the charts need to be modified to not treat those series as actual numerical values. 

Ex. Originally, the data for the chart looked like {x: test, alpha: 5, beta: 3}, and now it looks like {x: test, alpha: 5, beta: 3, group: story1, label: displayLabel1}. If the chart isn't modified, an error will be caused by the string values "story1" and "displayLabel1".

Another issue is that when a column in the chart is clicked, a tooltip should be shown with the option of opening the breakdown charts. These charts display the sub-values that make up the main metric value. The current ColumnChart doesn't support drawing tooltips, which is why I created these new charts.

Alternatively, I could add flags to the current charts so the above features are supported when they are turned on? That should also work, it would just require modifying the current charts somewhat.

Patch-set: 12
Reviewer: Gerrit User 1285369 <1285369@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>