Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> Ben, PTAL.
> 
> This is surely not ready to land as-is; but want to get your feedback on a few points:
> 
> - I'm not sure if the `Static` decorator is the best solution for our needs here. I wrote
>   what I think should be the corresponding implementations at the bottom of the file,
>   but they look kind of fiddly and verbose. Ideally, I think, I would just like to write
>   something like the `GetReportTemplates` function and get that magically piped through
>   the dashboard code as needed.

Thanks for the feedback! It is possible that Static is not quite the right API, but I think there are some compelling reasons to keep some of the more verbose bits. Let me know what you think!
I tried to address the specific fiddly/verbose concerns below. Are there any other specific parts of Static that seemed fiddly/verbose?

> - For these static templates, could we skip providing a `template_id` (could reuse `name`
>   for that purpose) and `modified` attributes (git history already records that)?

Names need to be mutable and are for display purposes only. Identifiers need to be immutable and are used for caching. (Modified timestamps are used for cache-busting, so thank you for setting them!) We're still dealing with a huge problem where bot names are being used as identifiers and needed to be changed.

Also, I'm sorry for writing misleading tests, but the api requires ids to be ints. For Static templates, they need to be hard-coded numbers; you can generate them using something like int(random.random()*(2**31)). I could make the api allow ids to be strings and handle the cast to int in report_template.py, but it's important that ids convey a sense of "danger, don't change this". I'll fix the tests and add an assertion to Static().

Does that help?

> - On the other side, I think we *should* be required to provide both owners and some sort
>   of documentation.

Requiring documentation is a great idea! I'll go do that.
Owners are only used to control who can modify dynamic templates. If you just want to document points of contact, would it be ok to keep that in the documentation?

> - Finally, not sure what the right location for a file like this should be?

Good question! dashboard/dashboard/ is super crowded. It's not a model or an api. It isn't a utility so it doesn't belong in common, but it needs to be imported into models/report_template.py, so I think I would rather keep this in common until we have 2 more sets of static templates, then it would make sense to add dashboard/dashboard/report_templates/.

Patch-set: 1
Reviewer: Gerrit User 1115874 <1115874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>