rendering: Fix filtering events.
It is possible for a segment to be completely contained in another
segment, or overlap other segments. The rendering metric processes
the trace events within these segments multiple times, leading to
unexpected results. This fixes that by updating the logic to filter
out events, so that each event is processed only once.

BUG=chromium:868650

Change-Id: Ie85afd9861037efd1d24588bedb883c079ab4d54