[pinpoint] Add adjustable significance thresholds to compare.py.
The bigger the regression, the fewer repeats we need to confirm it.
Having adjustable thresholds allows us to tune down the repeat count
for larger regressions.

In a future change, we can plumb through the expected regression size
from perf dashboard to Pinpoint, to hint at what thresholds to use.

Bug: catapult:#3936
Bug: catapult:#4508
Change-Id: Id8d775401afdf1a4e98473506f1fd8ba9de5f113