Update patch set 1
Patch Set 1:

After discussing with Ned via Hangout Chat, here is the proposed plan:

- For this CL, use `visualViewPortX` and `visualViewPortY` equals 0 if we are in compatibility mode, and remove this failback.

- For my previous CL about `pageScaleFactor`, use 1 if we are in compatibility mode, and remove the failback.

The purpose for this is to be able to compare our results in a fair way. So that we won't be using, for example, different `visualViewPortX` value for different milestones, which makes the comparison between versions unfair.

Patch-set: 1
Tag: mailMessageId=<CAH=bbm-yaq0nC5faDhzurgk33Tpd=vuW045DrLT_UrVUVdvHRA@mail.gmail.com>