[pinpoint] Add `comparison_magnitude` field to /api/new.
The new `comparison_magnitude` field allows API users to hint at the
size of regression to look for. Smaller regressions require more
repeats to confidently reproduce. This helps Pinpoint ensure that
it runs enough repeats to catch small regressions, but not waste
hardware resources on running extra repeats for large regressions.

For functional comparisons, this is a number between 0 and 1
representing the change in failure rate. For performance
comparisons, this is a number representing the absolute
change in any single statistic, e.g. mean, median, etc.

Bug: catapult:#3936
Change-Id: I3404c62e789411c3bc7b1291ec8997f8cdc30fde