Update patch set 1
Patch Set 1:

Hi Ned,

After discussing with Juan, we feel that following the proposed resolution will be too time-consuming and it's just not worth the effort. The reason is that in order to inject different javascripts, we will need to get the `--compatibility_mode` flag in `browser_option` and pass it all the way down to the, for example, `tap.py` which involves many changes along the way. Thus we feel that it will not be worth the time. So I will abandon this CL at the moment.

Patch-set: 1
Reviewer: Gerrit User 1292864 <1292864@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0