Update patch set 3
Patch Set 3:

Just a note to say that the external and internal representations of the categorical data are probably temporary. I've implemented it now so that it is ready for the demo but will need to work on better formats as we discussed. Also, this is chained to the dot zoom cl due to changes to the graph data processing implementation there. Thanks :)

Patch-set: 3
Reviewer: Gerrit User 1175774 <1175774@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1293423 <1293423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false