Functional prototype for dotPlot and linePlot
The user is able to choose now from both the drop-down menu and the table.
The chosen items can be displayed both in cumulative frequency plot and
dot plot after choosing a specific metric, a specific story, a diagnostic and
some of the available sub-diagnostics. The choice of sub-diagnostics is not
restricted. It is possible to pick one or more of them and have a plot
with all of them to notice the difference.

Bug: chromium:866423
Change-Id: I9b5936123a7cfa59686a6028ce39766ed8085d2a