Functional prototype for dotPlot and linePlot
The user is able to choose now from both the drop-down menu and the table wich
metric, story and diagnostic to be analysed.
The chosen items can be displayed both in cumulative frequency plot and
dot plot after choosing a specific metric, a specific story, a diagnostic and
some of the available sub-diagnostics. The choice of sub-diagnostics is not
restricted. It is possible to pick one or more of them and have a plot
to notice the difference.
Regarding the table, this process will be done picking a row with the name of
one metric. Then all the stories from that metric will be displayed and the
user should choose one. After that, being given all the diagnostics and picking
one of them, the user will be provided with all available sub-diagnostics. In
this point, from the first choice until the last one, the plot will be updated
adding each component in the plot.

Bug: chromium:866423
Change-Id: I9b5936123a7cfa59686a6028ce39766ed8085d2a