Add support for ES modules in third_party/idb.
This chanage was necessary for writing unit tests in CL 1130105. In
that branch, Webpack is being used to minify/uglify all of the source
code for the Service Worker. The same code is also being imported from
the browser for tests via ECMAScript script module imports:

  <script type="module" src="..."></script>

Before this change, the browser would complain about idb not providing
an export named "default". With these changes, the browser has no
problem importing idb and Webpack is still happy.

In addition to this CL, I made an issue on idb's GitHub to add support
for ES modules. See https://github.com/jakearchibald/idb/issues/66

Bug: chromium:855252
Change-Id: Idfd27b05218221ab0b2f74c178f2520bc2216cea