Reland Configure gae_ts_mon in chromeperf dashboard.
This CL configures the chromeperf dashboard to report metrics using gae_ts_mon.

This CL was reverted due to a broken import in dispatcher.py:
https://chromium-review.googlesource.com/c/catapult/+/1170043

This CL adds a test to ensure that dispatcher.py can be imported.
Patchset 1 is the original CL plus a failing test that is fixed in Patchset 2.

Demo: https://dev-benjhayden-1173494-dot-chromeperf.appspot.com/

Change-Id: Id2718e5e6350814f1e7b063a3fcd045ef726c6c9