Remove /change_internal_only.
This form is used by admins to control which bots are internal only versus
externally available. It is very infrequently used: about once or twice per
quarter.
This CL removes this form in favor of allowing admins to call a function in the
dev console.

This is split from https://chromium-review.googlesource.com/c/catapult/+/1153585
Followup related tasks:
 * Remove bot_whitelist in favor of Bot entities' internal_only property.
 * Optimize performance of ChangeInternalOnly.

Change-Id: I1f839b2c91d42bde98e99829971559d9e912a353