Remove bot_whitelist.
There are several Bots whose internal_only doesn't match bot_whitelist. They are
all deprecated, so they don't need to be fixed, but having multiple sources of
the same information that aren't automatically updated can cause confusion.

This CL removes bot_whitelist so that the Bot entities are the sole source of truth
for internal_only.

Change-Id: Id98f74962fc37adc795cab297159d9d0c7b89937